scraper/requirements.txt
playwright
pinecone
requests
google-auth
python-dotenv